fix: Disentangle data and logic.
Breaking change:
  data_array has just name. no more:
  - full_name
  - array-id
  - action indices
  data_set has no more:
  - action_id_map

This allows the data_set and data_array to be more standalone, with the
loop adding information to them as needed.  The idea is to clean up also
data_array/set  API.

Data_set no longer maintains a map of param->array.  Dat_array
constructor is slimmed  down.  Loop is in charge of "labeling" the
arrays as setpoints.  Array id is now superfluous, but still in the
repr.